Add labelKFold to cross validation
Currently, only KFold is supported in cross validation. But in cases
when data is gathered from different subjects and we want to avoid
over-fitting. labelKFold is a variation of k-fold which ensures that
the same label is not in both testing and training sets.

Unit test -'test labelKFold', is also added in MLUtilsSuite

Signed-off-by: Vincent Xie <vincent.xie@intel.com>
Signed-off-by: VinceShieh <vincent.xie@intel.com>